requirements.txt
flask
flask-cors
groq
gunicorn
python-dotenv
tavily-python